Very quick refactor of the groups pages to get them themed like the rest of the app. no changes to inline styles yet.
